requirements.txt
faiss-cpu==1.7.4
sentence-transformers==2.7.0
numpy==1.26.4
scikit-learn==1.5.0
python-dateutil==2.9.0
tqdm==4.66.4
